Fixed implicit cast from double to size_t in statistics.
Review: https://reviews.apache.org/r/20614

+++ 3rdparty/libprocess/include/process/statistics.hpp
@@ -81,7 +81,7 @@ private:
 
     // Use linear interpolation.
     const double position = percentile * (values.size() - 1);
-    const size_t index = floor(position);
+    const size_t index = static_cast<size_t>(floor(position));
     const double delta = position - index;
 
     CHECK_GE(index, 0u);
